% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{memscore}
\alias{memscore-package}
\alias{memscore}
\title{memscore}
\description{
Native R wrappers for the \code{memscore} command line tools.
}
\details{
The package can either shell out to an external \code{memscore} CLI or call the
Python package in-process through \code{reticulate}.

Use \code{vignette("memscore", package = "memscore")} for a guided walkthrough.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/memscore/}
  \item \url{https://github.com/bbuchsbaum/memscore}
  \item Report bugs at \url{https://github.com/bbuchsbaum/memscore/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
