% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{memscore}
\alias{memscore-package}
\alias{memscore}
\title{memscore}
\description{
Native R wrappers for the \code{memscore} command line tools.
}
\details{
The package does not use \code{reticulate}. Instead, it shells out to an
external \code{memscore} CLI or to \verb{python -m memscore}.

Use \code{vignette("memscore", package = "memscore")} for a guided walkthrough.
}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
